Split slash command and functional modules
